% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShippingQuote.R
\name{DaysTillArrival}
\alias{DaysTillArrival}
\title{Title}
\usage{
DaysTillArrival(miles)
}
\arguments{
\item{miles}{A numeric vector used in calculation for days until package arrives}
}
\value{
number if business days
}
\description{
Title
}
