\name{Shipping}
\alias{UPScalc}
\title{Hello, World!}
\usage{
Volume_calculator()
ShippingCost()
packaging_cost()
DaysTillArrival()
}
\description{
Prints 'Gives shipping quote'.
}
\examples{
Volume_calculator(9,6,6)
ShippingCost(2000, 10, 500)
packaging_cost(2000, TRUE)
DaysTillArrival(2000)
}
