% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{packaging_cost}
\alias{packaging_cost}
\title{Title}
\usage{
packaging_cost(volume, fragile)
}
\arguments{
\item{volume}{A numeric vector used in calculation for packing and shipping cost}

\item{fragile}{A Boolean parameter used to help the cost of packaging}
}
\value{
cost of packing
}
\description{
Title
}
