% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShippingQuote.R
\name{ShippingCost}
\alias{ShippingCost}
\title{Title}
\usage{
ShippingCost(volume, weight, distance)
}
\arguments{
\item{volume}{A numeric vector used in calculation for packing and shipping cost}

\item{weight}{A numeric vector for calculation of shipping cost}

\item{distance}{A numeric vector for calculation of shipping cost}
}
\value{
cost of shipment
}
\description{
Title
}
