% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Volume_calculator}
\alias{Volume_calculator}
\title{Title}
\usage{
Volume_calculator(length, width, height)
}
\arguments{
\item{length}{one of three numeric measurements for calculation of volume}

\item{width}{one of three numeric measurements for calculation of volume}

\item{height}{one of three numeric measurements for calculation of volume}
}
\value{
volume of package
}
\description{
Title
}
